+++ R/extract_rules.R
@@ -1,5 +1,5 @@
 #' @title
-#' Extract Causal Rules
+#' Extract (Causal) Decision Rules
 #'
 #' @description
 #' Method for extracting causal rules from the Random Forest or Gradient
@@ -7,37 +7,20 @@
 #'
 #' @param treelist A list of decision trees.
 #' @param X The features matrix.
-#' @param ntrees The number of decision trees.
-#' @param ite_std The standardized ITE.
-#' @param take_1 Whether or not to call the take1 helper function.
-#' @param type_decay The type of decay to apply when pruning the rules.
+#' @param ntrees The number of (the first) decision trees considered.
+#' @param max_depth The number of top levels from each tree considered
+#' to extract conditions.
 #'
 #' @keywords internal
 #'
 #' @return
-#' A vector of causal rules.
+#' A vector of (Causal) Decision Rules.
 #'
-extract_rules <- function(treelist, X, ntrees, ite_std, take_1, type_decay) {
+extract_rules <- function(treelist, X, ntrees, max_depth) {
 
   rules <- inTrees::extractRules(treeList = treelist,
                                 X = X,
                                 ntree = ntrees,
-                                maxdepth = 15)
-  rules <- c(rules)
-
-  if (take_1) {
-    rules <- rules[take1(length(rules))]
-  }
-
-  rules_matrix <- matrix(rules)
-  colnames(rules_matrix) <- "condition"
-  metric <- inTrees::getRuleMetric(rules_matrix,
-                                  X,
-                                  ite_std)
-  pruned <- inTrees::pruneRule(metric,
-                              X,
-                              ite_std,
-                              0.025,
-                              typeDecay = type_decay)
-  return(unique(pruned[, 4]))
+                                maxdepth = max_depth)
+  return(rules)
 }