+++ functional_tests/ft_cre.R
@@ -1,0 +1,75 @@
+source("generate_cre_dataset.R")
+source("CRE.R")
+source('split_data.R')
+source('check_input_data.R')
+source("check_method_params.R")
+source("check_hyper_params.R")
+source('estimate_ite.R')
+source('estimate_ite_bart.R')
+source('estimate_ite_bcf.R')
+source('estimate_ite_sipw.R')
+source('estimate_ite_poisson.R')
+source('estimate_ite_ipw.R')
+source('estimate_ite_aipw.R')
+source('estimate_ite_cf.R')
+source('estimate_ite_oreg.R')
+source('estimate_ps.R')
+source("generate_rules.R")
+source("extract_rules.R")
+source("generate_rules_matrix.R")
+source("select_causal_rules.R")
+source("interpret_select_rules.R")
+source("estimate_cate.R")
+source("print.R")
+source("plot.R")
+source("generate_causal_rules.R")
+source("prune_rules.R")
+library("SuperLearner")
+
+# Generate sample data
+set.seed(2021)
+dataset_cont <- generate_cre_dataset(n = 300,
+                                     rho = 0,
+                                     n_rules = 2,
+                                     p = 10,
+                                     effect_size = 2,
+                                     binary = FALSE)
+
+y <- dataset_cont[["y"]]
+z <- dataset_cont[["z"]]
+X <- as.data.frame(dataset_cont[["X"]])
+X_names <- names(as.data.frame(X))
+
+method_params = list(ratio_dis = 0.25,
+                     ite_method_dis = "bart",
+                     include_ps_dis = TRUE,
+                     ps_method_dis = "SL.xgboost",
+                     ps_method_inf = "SL.xgboost",
+                     oreg_method_dis = "SL.xgboost",
+                     oreg_method_inf = "SL.xgboost",
+                     ite_method_inf = "bart",
+                     include_ps_inf = TRUE,
+                     include_offset = FALSE,
+                     cate_method = "DRLearner",
+                     cate_SL_library = "SL.xgboost",
+                     filter_cate = FALSE,
+                     offset_name = NA,
+                     random_state = 3591)
+
+hyper_params = list(effect_modifiers = c(),
+                    ntrees_rf = 100,
+                    ntrees_gbm = 50,
+                    node_size = 20,
+                    max_nodes = 5,
+                    max_depth = 15,
+                    max_decay = 0.025,
+                    type_decay = 2,
+                    t = 0.025,
+                    q = 0.8,
+                    stability_selection = FALSE,
+                    pfer_val=0.1)
+
+
+cre_results <- cre(y, z, X, method_params, hyper_params)
+summary(cre_results, method_params, hyper_params)
+plot(cre_results)