+++ tests/testthat/test-estimate_cate.R
@@ -20,6 +20,9 @@ test_that("CATE (DRLearner) Estimation Runs Correctly", {
   ntrees_gbm <- 50
   node_size <- 20
   max_nodes <- 5
+  max_depth <- 15
+  max_decay <- 0.025
+  type_decay <- 2
   t <- 0.025
   q <- 0.8
   pfer_val <- 0.1
@@ -64,10 +67,13 @@ test_that("CATE (DRLearner) Estimation Runs Correctly", {
 
   # Step 3: Generate rules list
   initial_rules_dis <- generate_rules(X_dis, ite_std_dis, ntrees_rf, ntrees_gbm,
-                                      node_size, max_nodes, random_state = 981)
+                                      node_size, max_nodes, max_depth,
+                                      random_state = 981)
+
+  rules_dis <- prune_rules(initial_rules_dis, X_dis, ite_std_dis, max_decay, type_decay)
 
   # Step 4: Generate rules matrix
-  rules_all_dis <- generate_rules_matrix(X_dis, initial_rules_dis, t)
+  rules_all_dis <- generate_rules_matrix(X_dis, rules_dis, t)
   rules_matrix_dis <- rules_all_dis[["rules_matrix"]]
   rules_matrix_std_dis <- rules_all_dis[["rules_matrix_std"]]
   rules_list_dis <- rules_all_dis[["rules_list"]]
@@ -270,6 +276,9 @@ test_that("CATE (cf-means) Estimation Runs Correctly", {
   ntrees_gbm <- 50
   node_size <- 20
   max_nodes <- 5
+  max_depth <- 15
+  max_decay <- 0.025
+  type_decay <- 2
   t <- 0.025
   q <- 0.8
   pfer_val <- 0.1
@@ -309,10 +318,15 @@ test_that("CATE (cf-means) Estimation Runs Correctly", {
   ite_std_dis <- ite_list_dis[["ite_std"]]
 
   # Generate rules list
-  initial_rules_dis <- CRE:::generate_rules(X_dis, ite_std_dis, ntrees_rf, ntrees_gbm,
-                                            node_size, max_nodes, random_state = 214)
+  initial_rules_dis <- CRE:::generate_rules(X_dis, ite_std_dis, ntrees_rf,
+                                            ntrees_gbm, node_size, max_nodes,
+                                            max_depth, random_state = 214)
+
+  rules_dis <- CRE:::prune_rules(initial_rules_dis, X_dis, ite_std_dis, max_decay, type_decay)
+
+
   # Generate rules matrix
-  rules_all_dis <- CRE:::generate_rules_matrix(X_dis, initial_rules_dis, t)
+  rules_all_dis <- CRE:::generate_rules_matrix(X_dis, rules_dis, t)
   rules_matrix_dis <- rules_all_dis[["rules_matrix"]]
   rules_matrix_std_dis <- rules_all_dis[["rules_matrix_std"]]
   rules_list_dis <- rules_all_dis[["rules_list"]]