+++ tests/testthat/test-generate_causal_rules.R
@@ -29,6 +29,9 @@ test_that("generate_causal_rules works as expected!", {
                       ntrees_gbm = 50,
                       node_size = 20,
                       max_nodes = 5,
+                      max_depth = 15,
+                      max_decay = 0.025,
+                      type_decay = 2,
                       t = 0.025,
                       q = 0.8,
                       stability_selection = TRUE,
@@ -55,9 +58,9 @@ test_that("generate_causal_rules works as expected!", {
 
   # Generate Causal Decision Rules
   select_rules <- generate_causal_rules(X, ite_std, method_params, hyper_params)
-  expect_true(class(select_rules) == "character")
+  expect_true(class(select_rules[[1]]) == "character")
 
   hyper_params[["effect_modifiers"]] <- X_names[c(5,7,8,9)]
   select_rules <- generate_causal_rules(X, ite_std, method_params, hyper_params)
-  expect_true(class(select_rules) == "character")
+  expect_true(class(select_rules[[1]]) == "character")
 })